fpdf: add magic bytes close to header
Ensure proper behavior of applications that look for magic bytes near the beginning of the file.

see 2008 PDF Spec 7.5.2 File Header

> If a PDF file contains binary data, as most do (see 7.2, "Lexical Conventions"), the header line shall be
immediately followed by a comment line containing at least four binary characters—that is, characters whose
codes are 128 or greater. This ensures proper behaviour of file transfer applications that inspect data near the
beginning of a file to determine whether to treat the file’s contents as text or as binary.

Reviewed-on: https://codeberg.org/go-pdf/fpdf/pulls/99
Co-authored-by: mspindelhirn <moritz.spindelhirn@timbertom.de>
Co-committed-by: mspindelhirn <moritz.spindelhirn@timbertom.de>